refactor: moving saved metrics pre conditions to the right code module.
